LPD-35630 Merge AM contributor modules
