+++ modules/apps/adaptive-media/adaptive-media-editor-configuration/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Adaptive Media Editor Configuration
+Bundle-SymbolicName: com.liferay.adaptive.media.editor.configuration
+Bundle-Version: 1.0.0
+-dsannotations-options: inherit
