+++ modules/apps/adaptive-media/adaptive-media-editor-configuration/build.gradle
@@ -1,0 +1,18 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.1"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
+	compileOnly project(":apps:adaptive-media:adaptive-media-image-api")
+	compileOnly project(":apps:adaptive-media:adaptive-media-image-item-selector-api")
+	compileOnly project(":apps:blogs:blogs-api")
+	compileOnly project(":apps:blogs:blogs-item-selector-api")
+	compileOnly project(":apps:item-selector:item-selector-api")
+	compileOnly project(":apps:item-selector:item-selector-criteria-api")
+	compileOnly project(":apps:journal:journal-api")
+	compileOnly project(":apps:journal:journal-item-selector-api")
+	compileOnly project(":core:petra:petra-function")
+
+	compileOnly project(":core:petra:petra-string")
+	testImplementation group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
+	testImplementation group: "org.skyscreamer", name: "jsonassert", version: "1.2.3"
+}
